synapse-to-sharepoint-complete/synapse-to-sharepoint-complete/synapse-sharepoint-seplat/requirements.txt
# Copy from main project requirements
# Run from parent directory: pip install -r ../synapse-to-sharepoint/requirements.txt

# Or use this file directly:
azure-functions==1.18.0
azure-identity==1.15.0
pyodbc==5.0.1
msal==1.26.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
structlog==24.1.0
